pyproject.toml
[project]
name = "llm-assistant-backend"
version = "0.3.0"
description = "Backend service for multifunctional LLM assistant"
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    # API
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.30.0",
    "python-dotenv==1.0.1",
    # PT Portal loader / utils
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.2.1,<3",
    "PyYAML~=6.0.1",
    "tqdm~=4.66.0",
    "requests>=2.31.0,<3",
    "urllib3>=2.2.0,<3",
    "requests-cache>=1.2.0,<2",
    "beautifulsoup4>=4.12.2,<5",
    "lxml>=5.2.0,<6",
    "readability-lxml~=0.8.1",
    "pdfminer.six>=20231228,<20240000",
    "python-slugify~=8.0.4",
    "rapidfuzz>=3.9.6,<4",
    "tiktoken>=0.7,<1",
    "regex>=2023.12.25,<2026",
    # Embeddings & Stores
    "numpy>=1.26,<2",
    "sentence-transformers>=2.7.0,<3",
    "openai>=1.43.0,<2",
    "qdrant-client>=1.9.1,<2",
    "elasticsearch>=8.13.0,<9",
    "httpx>=0.27.0,<1.0",
    "tenacity>=8.2.0,<9.0",
    # PostgreSQL-only NL→SQL
    "SQLAlchemy>=2.0.25,<3",
    "psycopg[binary]>=3.1.12,<4", # бинарные колеса; для прод можно заменить на psycopg[c]
]

[project.optional-dependencies]
local = [
    "transformers>=4.37.0,<5",
    "torch>=2.1.0,<3",
    "llama-cpp-python>=0.2.0,<0.3",
    "accelerate>=0.27.0,<1.0",
]

dev = [
    "pytest>=8.2,<9",
    "pytest-cov>=5,<6",
    "ruff>=0.5,<1", # опционально для стиля
    "mypy>=1.10,<2", # опционально для типизации
]

[project.scripts]
pt-loader = "tools.rag.sources.pt_portal.cli:main"
rag-chunk = "tools.rag.chunking.cli:main"
rag-embed = "tools.rag.embeddings.cli:main"
sql-seed = "tools.sql.seed_data:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# В wheel должны попасть и FastAPI-приложение, и тулзы из tools/*
packages = ["apps", "tools"]

[tool.uv]
index-url = "https://pypi.org/simple"